+++ packages/@expo/metro-runtime/async-require.js
@@ -1,1 +1,4 @@
-require('./src/async-require');
+console.warn(
+  '@expo/metro-runtime/async-require is deprecated. All projects have async require support by default.'
+);
+// require('./src/async-require');